Restrict inner function.
Add a restriction on inner function in fusion such that
it can only have constructor tails. This is needed to
keep the invariant that each function is a single
to level case expression.

Namely non-constuctor tails results in nested cases
expresison in the current implementaiton.

The other approach to handle those cases is to allow
such cases to be folded into outer call.

However for now, the current handling is suficient.