Eliminate redundant inputs during tupling
while tupling a sequence of calls if the some passed
arguments have the same value then those are identified and
passed once. That fact is used to optimize the body of the
tupled funcion.

This update solve a tupling terminiation issue for some input
cases and it solves a performance issue for other.