L4.Tail: add EndOfMain marker; fix codegen to not generate two returns
We already always add a printing statement to the main expression, so
the main expression is already special. Now we have a dedicated marker
for the end of main, so that we can add some more specialty in the
codegen. In particular, we don't generate a return statement for main in
the generic codegen pass (`codegenTail`). Instead, a shutdown code for
the whole program is generated elsewhere. This has been done before,
but the main pass generated an extra return for main (because it looked
like a regular function) and the shutdown code was never executed. The
new special marker allows to tell between a regular function and main,
and to skip the return statement in the latter case.

Fix #242